test_conversation_handler.py
```python
import unittest
from unittest.mock import patch
from conversation_handler import ConversationHandler

class TestConversationHandler(unittest.TestCase):
    def setUp(self):
        self.conversation_handler = ConversationHandler()
        self.sample_content = [
            {
                'title': 'AI in Healthcare',
                'summary': 'This article discusses the impact of AI in healthcare...'
            },
            {
                'title': 'The Future of AI',
                'summary': 'This article explores what the future holds for AI...'
            }
        ]

    def test_start_conversation(self):
        with patch('builtins.input', return_value='Interesting article!'):
            self.conversation_handler.start_conversation(self.sample_content)
            self.assertEqual(len(self.conversation_handler.get_conversation_history()), 2)

    def test_discuss_content(self):
        with patch('builtins.input', return_value='Great read!'):
            self.conversation_handler.discuss_content(self.sample_content[0])
            self.assertEqual(len(self.conversation_handler.get_conversation_history()), 1)

    def test_get_conversation_history(self):
        with patch('builtins.input', return_value='Very informative!'):
            self.conversation_handler.start_conversation(self.sample_content)
            conversation_history = self.conversation_handler.get_conversation_history()
            self.assertEqual(len(conversation_history), 2)
            self.assertEqual(conversation_history[0]['content'], self.sample_content[0])
            self.assertEqual(conversation_history[0]['user_feedback'], 'Very informative!')

    def test_get_user_feedback(self):
        with patch('builtins.input', return_value='I learned a lot!'):
            self.conversation_handler.start_conversation(self.sample_content)
            user_feedback = self.conversation_handler.get_user_feedback()
            self.assertEqual(len(user_feedback), 2)
            self.assertEqual(user_feedback[0], 'I learned a lot!')

if __name__ == '__main__':
    unittest.main()
```
